Wrapperの大分離
