+++ src/components/Components.d.ts
@@ -4,3 +4,7 @@ export interface Stylable {
   className?: string
   css?: InterpolationWithTheme<any>
 }
+
+export interface WithClassName {
+  className?: string
+}
