+++ src/components/templates/templates.d.ts
@@ -1,0 +1,8 @@
+import { Stylable } from "../Components"
+
+export type BackgroundProps = Stylable & Meta
+
+export interface Meta {
+  description: string
+  title: string
+}
